# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A secure file storage REST API with S3 and CloudFront",
        "title": "File Vault API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/mohamed8eo/file-vault",
            "email": "support@filevault.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": ["http", "https"],
    "paths": {
        "/auth/sign-up": {
            "post": {
                "tags": ["Authentication"],
                "summary": "Register new user",
                "description": "Register a new user with name, email and password",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "user",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "name": {"type": "string"},
                            "email": {"type": "string"},
                            "password": {"type": "string"}
                        }
                    }
                }],
                "responses": {
                    "200": {"description": "User registered successfully"},
                    "400": {"description": "Invalid request"},
                    "409": {"description": "User already exists"}
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": ["Authentication"],
                "summary": "Login user",
                "description": "Authenticate user and get access/refresh tokens",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "credentials",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "email": {"type": "string"},
                            "password": {"type": "string"}
                        }
                    }
                }],
                "responses": {
                    "200": {"description": "Login successful"},
                    "400": {"description": "Invalid request"},
                    "401": {"description": "Invalid credentials"}
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": ["Authentication"],
                "summary": "Refresh access token",
                "description": "Get new access token using refresh token cookie",
                "produces": ["application/json"],
                "responses": {
                    "200": {"description": "Token refreshed"},
                    "401": {"description": "Invalid or expired refresh token"}
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization",
            "description": "Type 'Bearer' followed by a space and JWT token."
        }
    }
}